Moved logger to `AbstractClient`
