+++ tests/EchoLogger.php
@@ -20,5 +20,9 @@ class EchoLogger extends AbstractLogger
     public function log($level, $message, array $context = array())
     {
         echo sprintf('[%s] %s: %s' . PHP_EOL, date('Y-m-d H:i:s'), strtoupper($level), $message);
+        if (!empty($context)) {
+            print_r($context);
+            echo PHP_EOL;
+        }
     }
 }
